backend/requirements.txt
# Smart Reading List Backend

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.26.0

# Database
aiosqlite>=0.19.0

# Vector store - use chromadb-client for lighter install
chromadb>=0.4.22

# HTML parsing
trafilatura>=1.6.0
readability-lxml>=0.8.1

# For ONNX runtime (required by chromadb)
onnxruntime>=1.14.1
